Worked on the all other module to applying the permissions

+++ src/app/features/subjects/pages/subject-view/subject-view.component.ts
@@ -4,6 +4,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { MaterialModule } from '../../../../shared/modules/material/material.module';
 import { SubjectService } from '../../services/subject.service';
 import { ErrorHandlerService } from '../../../../core/services/error-handler.service';
+import { PermissionService } from '../../../../core/services/permission.service';
 import { Subject } from '../../../../core/models/subject.model';
 
 @Component({
@@ -18,12 +19,20 @@ export class SubjectViewComponent implements OnInit {
   isLoading = true;
   subjectId!: number;
 
+  // Permission checks
+  hasEditPermission = false;
+  hasDeletePermission = false;
+
   constructor(
     private subjectService: SubjectService,
     private route: ActivatedRoute,
     private router: Router,
-    private errorHandler: ErrorHandlerService
-  ) {}
+    private errorHandler: ErrorHandlerService,
+    private permissionService: PermissionService
+  ) {
+    this.hasEditPermission = this.permissionService.hasPermission('subjects.edit');
+    this.hasDeletePermission = this.permissionService.hasPermission('subjects.delete');
+  }
 
   ngOnInit(): void {
     this.route.params.subscribe(params => {
@@ -53,10 +62,19 @@ export class SubjectViewComponent implements OnInit {
   }
 
   onEdit(): void {
+    if (!this.hasEditPermission) {
+      this.errorHandler.showError('You do not have permission to edit subjects');
+      return;
+    }
     this.router.navigate(['/subjects/edit', this.subjectId]);
   }
 
   onDelete(): void {
+    if (!this.hasDeletePermission) {
+      this.errorHandler.showError('You do not have permission to delete subjects');
+      return;
+    }
+    
     if (confirm(`Are you sure you want to delete subject "${this.subject?.name}"?`)) {
       this.subjectService.deleteSubject(this.subjectId).subscribe({
         next: (response: any) => {
